Cleanup and refactoring of yahoo discovery service
